Edited inline CSS for the Google map; moved it to the external CSS file.
The reason the map was not appearing is that the browser didn't like the use of percentages in the context of its parent element (probably because the parent element, <body>, didn't have a defined width and height). I changed the CSS rule to use pixel values. See lines 72-79. I also cleaned up the readability of the CSS file.